moved entites from common directory
