+++ libs/common/src/entities/index.ts
@@ -1,3 +1,3 @@
 export * from './user.entity';
-export * from './chat.entity';
-export * from './message.entity';
+export * from '../../../../src/entities/chat.entity';
+export * from '../../../../src/entities/message.entity';
