+++ src/entities/chat.entity.ts
@@ -7,7 +7,7 @@ import {
   PrimaryGeneratedColumn,
 } from 'typeorm';
 import { Message } from './message.entity';
-import { User } from './user.entity';
+import { User } from '../../libs/common/src/entities/user.entity';
 @Entity('Chat')
 export class Chat {
   @PrimaryGeneratedColumn()
