+++ src/entities/message.entity.ts
